Split RabbitMQ configuration between producer and listener
