+++ pet-hub-api/src/main/java/com/pet/pethubapi/infrastructure/rabbitmq/RabbitMqProducerConfiguration.java
@@ -2,12 +2,9 @@ package com.pet.pethubapi.infrastructure.rabbitmq;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.AmqpException;
-import org.springframework.amqp.rabbit.annotation.EnableRabbit;
-import org.springframework.amqp.rabbit.config.SimpleRabbitListenerContainerFactory;
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
-import org.springframework.amqp.support.converter.MessageConverter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.retry.RetryCallback;
@@ -19,29 +16,12 @@ import java.time.Duration;
 
 @Slf4j
 @Configuration
-@EnableRabbit
-public class RabbitMqConfiguration {
-
-    /**
-     * Configuration does not implement Spring retry policy due to the fact it has blocking effect on consumer (each retry keeps consumer instance blocked until all retries are exhausted).
-     * Having concurrent consumers could mitigate that issue, but still is blocking. Implemented solution is using retry queue combined with death-letter exchange and parking-lot queue, which is not blocking consumer instance.
-     */
-    @Bean
-    SimpleRabbitListenerContainerFactory rabbitListenerContainerFactory(final ConnectionFactory connectionFactory) {
-        final var factory = new SimpleRabbitListenerContainerFactory();
-        factory.setConnectionFactory(connectionFactory);
-        factory.setMessageConverter(rabbitJsonMessageConverter());
-        factory.setPrefetchCount(0); // one consumer will not prefetch more than one message at a time (default 250 prefetched messages in memory; configured in favor of concurrent consumers)
-        factory.setConcurrentConsumers(3); // defines number of concurrent consumers (scenario where producer sends multiple messages at once to same consumer; avoid storing queued messages in memory)
-        factory.setMaxConcurrentConsumers(5);
-        factory.setDefaultRequeueRejected(false);
-        return factory;
-    }
+public class RabbitMqProducerConfiguration {
 
     @Bean
     RabbitTemplate rabbitTemplate(final ConnectionFactory connectionFactory) {
         final var template = new RabbitTemplate(connectionFactory);
-        template.setMessageConverter(rabbitJsonMessageConverter());
+        template.setMessageConverter(new Jackson2JsonMessageConverter());
         template.setObservationEnabled(true);
 
         final var retryPolicy = new RetryTemplateBuilder()
@@ -61,11 +41,6 @@ public class RabbitMqConfiguration {
             .build();
         template.setRetryTemplate(retryPolicy);
         return template;
-    }
-
-    @Bean
-    MessageConverter rabbitJsonMessageConverter() {
-        return new Jackson2JsonMessageConverter();
     }
 
 }
