+++ pet-hub-api/src/main/resources/application.properties
@@ -61,9 +61,9 @@ management.endpoint.health.group.pet-hub-api.show-details=when_authorized
 # Swagger-UI
 springdoc.swagger-ui.path=/swagger-ui.html
 # custom
+pet.rabbitmq.enabled=false
 pet.jwt.secret-key=W3BldGh1YjIwMjRdVGhpcy5pcy52ZXJ5LnNlY3VyZS5zaWduLmtleTMyIQ==
 ## in milliseconds
 pet.jwt.access-expiration=3600000
 pet.jwt.admin-access-expiration=3600000
 pet.jwt.refresh-expiration=86400000
-pet.rabbitmq.enabled=true
