+++ pet-hub-batch/src/main/resources/application.properties
@@ -35,7 +35,13 @@ spring.jpa.properties.hibernate.connection.provider_disables_autocommit=true
 spring.jpa.properties.hibernate.query.in_clause_parameter_padding=true
 spring.jpa.properties.hibernate.query.fail_on_pagination_over_collection_fetch=true
 spring.jpa.properties.hibernate.query.plan_cache_max_size=4096
-# Actuator
+# RabbitMQ
+spring.rabbitmq.host=localhost
+spring.rabbitmq.port=5672
+spring.rabbitmq.username=pet-batch
+spring.rabbitmq.password=pet-batch-password
+spring.rabbitmq.virtual-host=pet-hub
+# Actuator (api/actuator/info & api/actuator/health/pet-hub-batch)
 management.endpoints.web.exposure.include=health,info,metrics,prometheus
 management.metrics.tags.application-name=${spring.application.name}
 management.endpoint.health.probes.enabled=true
@@ -55,6 +61,7 @@ management.endpoint.health.group.pet-hub-batch.show-details=when_authorized
 # Swagger-UI
 springdoc.swagger-ui.path=/swagger-ui.html
 # custom
+pet.rabbitmq.enabled=true
 pet.jwt.secret-key=W3BldGh1YjIwMjRdVGhpcy5pcy52ZXJ5LnNlY3VyZS5zaWduLmtleTMyIQ==
 ## in milliseconds
 pet.jwt.access-expiration=3600000
